[Refactor] proceduredetail로 패키징
