+++ app/src/main/AndroidManifest.xml
@@ -25,7 +25,7 @@
             android:name=".ui.budget.procedurecontent.ProcedureContentActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.budget.ProcedureDetailActivity"
+            android:name=".ui.budget.proceduredetail.ProcedureDetailActivity"
             android:exported="false" />
         <activity
             android:name=".ui.budget.budgetdetail.main.BudgetDetailActivity"
