+++ app/src/main/java/kr/sparta/tripmate/ui/viewmodel/budget/proceduredetail/ProcedureDetailViewModel.kt
@@ -1,4 +1,4 @@
-package kr.sparta.tripmate.ui.viewmodel.budget
+package kr.sparta.tripmate.ui.viewmodel.budget.proceduredetail
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
