+++ app/src/main/res/layout/activity_procedure_detail.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".ui.budget.ProcedureDetailActivity">
+    tools:context=".ui.budget.proceduredetail.ProcedureDetailActivity">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/procedure_toolbar"
